+++ src/export/tex/types.ts
@@ -1,0 +1,6 @@
+export interface TexExportOptions {
+  filename: string;
+  images?: string;
+  template?: string;
+  options?: string;
+}
